+++ Examples/Subscriptions/src/Internal/TimestampedValue.hpp
