+++ README.md
@@ -23,7 +23,7 @@ This project aims to supply, over time, examples and best practices of PLCnext T
 |[03](Examples/CppDataTypeTest/)| [Data Type Exchange Example](Examples/CppDataTypeTest/README.md)| This example shows how to transfer data from a C++ project to the GDS by declaring *Port variables* in C++. Variables published to the GDS can be accessed by other programs (IEC/C++/Simulink/C#) or the IOs| [SourceCode](Examples/CppDataTypeTest/src/)
 |[04](Examples/ThreadExample/)| [Thread Example](Examples/ThreadExample/README.md)| This describes how to implement Threads in a component of C++ project.| [SourceCode](Examples/ThreadExample/src/)
 |[05](Examples/DataAccess/)| [Data Access Example](Examples/DataAccess/README.MD)| This example shows how to read and write GDS variables using the Data Access service.| [SourceCode](Examples/DataAccess/src/)
-|[06](Examples/SubscriptionExample/)| [Subscription Example](Examples/SubscriptionExample/README.MD)| This example shows how to read GDS variables using the Subscription service.| [SourceCode](Examples/SubscriptionExample/src/)
+|[06](Examples/Subscriptions/)| [Subscription Example](Examples/Subscriptions/README.MD)| This example shows how to read GDS variables using the Subscription service.| [SourceCode](Examples/Subscriptions/src/)
 |[07](Examples/Force/)| [Force Example](Examples/Force/README.MD)| This example shows how GDS variables can be forced using the Force service.| [SourceCode](Examples/Force/src/)
 |[08](Examples/NotificationExample/)| [Notification Example](Examples/NotificationExample/README.MD)| This example shows how to use the Notification Manager.| [SourceCode](Examples/NotificationExample/src/)
 |[09](Examples/IncludeOpenSourceLibrary/)| [Include an Open Source Library](Examples/IncludeOpenSourceLibrary/README.md)| This describes how to use a third-party open-source library in a C++ project that is built for PLCnext Control using Eclipse|
